updated middleware namespace
