+++ src/app/Middleware/RequestLogger/loggerOptions.cs
@@ -1,0 +1,14 @@
+﻿namespace Middleware
+{
+    /// <summary>
+    /// Logger options used to configure DI
+    /// </summary>
+    public class LoggerOptions
+    {
+        public bool Log2xx { get; set; } = true;
+        public bool Log3xx { get; set; } = true;
+        public bool Log4xx { get; set; } = true;
+        public bool Log5xx { get; set; } = true;
+        public double TargetMs { get; set; } = 250;
+    }
+}
